hershey-font-check: check for zero-width glitches in .jhf files
A glitch (a glyph with left == right bounds, i.e zero-width) has been
observed in the capital-J glyph (3010) in some versions of some
hersheyfonts .jhf file distributions.

This new check in hershey-font-check now catches the one instance of
this glitch in this distribution (the capital-J in rowmant.jhf).

Note that there actually are a few glyphs (in symbolic.jhf) which are
properly zero-width, as they are purely vertical lines.  This check
recognizes that case.

Signed-off-by: Kamal Mostafa <kamal@whence.com>